+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - added support for defining supported OS in the extension and showing an info if the extension is not supported on the current OS
 - added support for using external libraries for extension UI components
 - added support for new UI extension placements ('tasks-sidebar-actions-left', 'tasks-sidebar-actions-right') to be used in Tasks header
+- added activateTask action to extension UI component props
 
 ## [0.67.0]